Mentioned the Sandbox folder.
The Sandbox folder serves the same purpose as any Bundles folder, but it made specifically for custom content. It managed to disappear a long time ago from the install of Unturned. I remembered that in 2017 it did exist, and an issue was created that brought back this folder to the install: https://github.com/SmartlyDressedGames/Unturned-3.x-Community/issues/3080

Signed-off-by: Jdance-Media <67397422+Jdance-Media@users.noreply.github.com>